pyproject.toml
[project]
name = "oreilly-reasoning-models"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ace-tools>=0.0",
    "anthropic>=0.57.1",
    "httpx>=0.28.1",
    "ipykernel>=6.29.5",
    "ipython>=9.4.0",
    "matplotlib>=3.10.3",
    "numpy>=2.3.1",
    "ollama>=0.5.1",
    "openai>=1.93.3",
    "pandas>=2.3.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "seaborn>=0.13.2",
    "tabulate>=0.9.0",
]
